Fixed a few warnings.
Rev: src/modules/Mird/mird_glue.c:1.8
Rev: src/modules/_Crypto/lib/md2.c:1.3

+++ src/modules/_Crypto/lib/md2.c
@@ -1,6 +1,6 @@
 
 /*
- * $Id: md2.c,v 1.2 2000/08/02 18:13:51 grubba Exp $
+ * $Id: md2.c,v 1.3 2000/08/13 14:57:24 grubba Exp $
  *
  *  md2.c : MD2 hash algorithm.
  *
@@ -54,7 +54,8 @@ void md2_update(struct md2_ctx *ctx,
   unsigned INT32 L;
   while (len) 
   {
-    L = (16 < (len + ctx->count)) ? (16-ctx->count) : len;
+    L = (((unsigned INT32)16) < (len + ctx->count)) ?
+      ((unsigned INT32)(16-ctx->count)) : len;
     memcpy(ctx->buf+ctx->count, buffer, L);
     ctx->count+=L;
     buffer+=L;
